Add support for looking up rate adjust mods when calculating play time
